login and middleware

+++ routes/web.php
@@ -18,25 +18,31 @@
     // Define the route to render the "products" view
     Route::get('/', [ProductController::class, 'index']);
 
-    Route::get('/product/create', [ProductController::class, 'create']);
+    Route::get('/product/create', [ProductController::class, 'create'])->middleware(('auth'));
 
     Route::get('/product/{product}', [ProductController::class, 'show']);
 
-    Route::post('/product', [ProductController::class, 'store']);
+    Route::post('/product', [ProductController::class, 'store'])->middleware(('auth'));
 
-    Route::get('/products/{product}/edit', [ProductController::class, 'edit']);
+    Route::get('/products/{product}/edit', [ProductController::class, 'edit'])->middleware(('auth'));
 
-    Route::put('/products/{product}', [ProductController::class, 'update']);
+    Route::put('/products/{product}', [ProductController::class, 'update'])->middleware(('auth'));
     
-    Route::delete('/products/{product}', [ProductController::class, 'destroy']);
+    Route::delete('/products/{product}', [ProductController::class, 'destroy'])->middleware(('auth'));
     
-    Route::get('/register', [UserController::class, 'create']);
+    Route::get('/register', [UserController::class, 'create'])->middleware(('guest'));
 
     //save new user
-    Route::post('/users', [UserController::class, 'store']);
+    Route::post('/users', [UserController::class, 'store'])->middleware(('guest'));
 
     //logout
-    Route::post('/logout', [UserController::class,'logout']);
+    Route::post('/logout', [UserController::class,'logout'])->middleware(('auth'));
+
+    //Login
+    Route::get('/login', [UserController::class, 'login'])->name('login')->middleware(('guest'));
+
+    //authenticate login form
+    Route::post('/users/authenticate', [UserController:: class, 'authenticate'])->middleware(('guest'));
     /*
     index - show all products
     show - show single product
